+++ src/App.jsx
@@ -1,4 +1,8 @@
-import React, { useState, useEffect, useRef, useMemo, useCallback, memo } from 'react';
+import React, { useState, useEffect, useRef, useMemo, useCallback } from 'react';
+import TuringProgram from './TuringProgram.js';
+import ColorPalette from './ColorPalette.jsx';
+import TransitionTable from './TransitionTable.jsx';
+import Controls from './Controls.jsx';
 
 function hslToRgb(h, s, l) {
   let r, g, b;
@@ -26,15 +30,6 @@ function hslToRgb(h, s, l) {
   return [Math.round(r * 255), Math.round(g * 255), Math.round(b * 255)];
 }
 
-const DIRECTIONS = {
-  LEFT: 0,
-  RIGHT: 1,
-  UP: 2,
-  DOWN: 3
-};
-
-const DIRECTION_NAMES = ['LEFT', 'RIGHT', 'UP', 'DOWN'];
-
 const generateColorPalette = (numSymbols) => {
   const colors = [];
     
@@ -51,223 +46,6 @@ const generateColorPalette = (numSymbols) => {
   return colors;
 };
 
-const createRandom = (seed) => {
-  let state = seed;
-  return () => {
-    state = (state * 1664525 + 1013904223) % 4294967296;
-    return state / 4294967296;
-  };
-};
-
-class TuringProgram {
-  constructor(numStates, numSymbols, width, height, numHeads, headRadius, seed, transitionTable = null, initialHeads = null) {
-    this.numStates = numStates;
-    this.numSymbols = numSymbols;
-    this.width = width;
-    this.height = height;
-    this.numHeads = numHeads;
-    this.headRadius = headRadius;
-    this.seed = seed;
-    this.initialHeads = initialHeads;
-    this.table = new Int32Array(numStates * numSymbols * 3);
-    this.grid = new Int32Array(width * height);
-    this.heads = new Array(numHeads);
-    this.iterations = 0;
-        
-    const random = createRandom(seed);    
-    
-    if (transitionTable) {
-      for (const transition of transitionTable) {
-        const state = transition.currentState;
-        const symbol = transition.currentSymbol;
-        const newState = transition.newState;
-        const newSymbol = transition.newSymbol;
-        const action = typeof transition.action === 'string' 
-          ? DIRECTION_NAMES.indexOf(transition.action)
-          : transition.action;
-          
-        const idx = (numStates * symbol + state) * 3;
-        this.table[idx] = newState;
-        this.table[idx + 1] = newSymbol;
-        this.table[idx + 2] = action;
-      }
-    } else {        
-      for (let state = 0; state < numStates; state++) {
-        for (let symbol = 0; symbol < numSymbols; symbol++) {          
-          const newState = Math.floor(random() * numStates);         
-          const newSymbol = (symbol === 0) 
-            ? 1 + Math.floor(random() * (numSymbols - 1))
-            : Math.floor(random() * numSymbols);
-          const action = Math.floor(random() * 4);
-          const idx = (numStates * symbol + state) * 3;
-          this.table[idx] = newState;
-          this.table[idx + 1] = newSymbol;
-          this.table[idx + 2] = action;
-        }
-      }
-    }
-    
-    this.reset();
-  }
-  
-  reset() {
-    this.grid.fill(0);
-          
-    const centerX = Math.floor(this.width / 2);
-    const centerY = Math.floor(this.height / 2);
-          
-    const radius = Math.min(this.width, this.height) * this.headRadius;
-    const random = createRandom(this.seed);
-    
-    for (let i = 0; i < this.numHeads; i++) {
-      const angle = (i * 2 * Math.PI) / this.numHeads;
-      const x = centerX + Math.round(radius * Math.cos(angle));
-      const y = centerY + Math.round(radius * Math.sin(angle));
-            
-      const boundedX = Math.max(0, Math.min(this.width - 1, x));
-      const boundedY = Math.max(0, Math.min(this.height - 1, y));
-            
-      if (!this.heads[i]) {
-        this.heads[i] = { state: 0, x: boundedX, y: boundedY };
-      } else {
-        this.heads[i].state = 0;
-        this.heads[i].x = boundedX;
-        this.heads[i].y = boundedY;
-      }
-              
-      // Ensure symbol is within bounds of numSymbols
-      this.grid[boundedY * this.width + boundedX] = 1 + (i % (this.numSymbols - 1));
-    }
-          
-    const patternSize = Math.min(5, Math.floor(Math.min(this.width, this.height) / 40));
-    for (let x = -patternSize; x <= patternSize; x++) {
-      for (let y = -patternSize; y <= patternSize; y++) {      
-        if (random() < 0.8) continue;
-        
-        const gx = (centerX + x + this.width) % this.width;
-        const gy = (centerY + y + this.height) % this.height;
-                  
-        // Ensure symbol is within bounds of numSymbols
-        this.grid[gy * this.width + gx] = 1 + Math.floor(random() * (this.numSymbols - 1));
-      }
-    }
-    
-    this.iterations = 0;
-  }
-  
-  update(steps) {
-    const width = this.width;
-    const numStates = this.numStates;
-    
-    for (let i = 0; i < steps; i++) {
-      for (let h = 0; h < this.heads.length; h++) {
-        const head = this.heads[h];
-        const gridIdx = head.y * width + head.x;
-        const symbol = this.grid[gridIdx];
-        const state = head.state;
-        const tableIdx = (numStates * symbol + state) * 3;
-        
-        head.state = this.table[tableIdx];
-        // Ensure new symbol is within bounds
-        const newSymbol = this.table[tableIdx + 1];
-        this.grid[gridIdx] = newSymbol < this.numSymbols ? newSymbol : 0;
-        
-        switch (this.table[tableIdx + 2]) {
-          case DIRECTIONS.LEFT:
-            head.x = (head.x + 1) % width;
-            break;
-            
-          case DIRECTIONS.RIGHT:
-            head.x = (head.x - 1 + width) % width;
-            break;
-            
-          case DIRECTIONS.UP:
-            head.y = (head.y - 1 + this.height) % this.height;
-            break;
-            
-          case DIRECTIONS.DOWN:
-            head.y = (head.y + 1) % this.height;
-            break;
-        }
-      }
-      
-      this.iterations++;
-    }
-  }
-  
-  exportToJson() {    
-    const transitionTable = [];
-    
-    for (let state = 0; state < this.numStates; state++) {
-      for (let symbol = 0; symbol < this.numSymbols; symbol++) {
-        const idx = (this.numStates * symbol + state) * 3;
-        const newState = this.table[idx];
-        const newSymbol = this.table[idx + 1];
-        const action = this.table[idx + 2];
-        
-        transitionTable.push({
-          currentState: state,
-          currentSymbol: symbol,
-          newState: newState,
-          newSymbol: newSymbol,
-          action: DIRECTION_NAMES[action]
-        });
-      }
-    }
-    
-    const headPositions = this.heads.map(head => ({
-      state: head.state,
-      x: head.x,
-      y: head.y
-    }));
-    
-    return {
-      parameters: {
-        numStates: this.numStates,
-        numSymbols: this.numSymbols,
-        numHeads: this.numHeads,
-        headRadius: this.headRadius,
-        width: this.width,
-        height: this.height,
-        seed: this.seed,
-        iterations: this.iterations
-      },
-      transitionTable: transitionTable,
-      currentHeads: headPositions
-    };
-  }
-  
-  getTransitionTableDisplay() {
-    const table = [];
-    
-    for (let state = 0; state < this.numStates; state++) {
-      for (let symbol = 0; symbol < this.numSymbols; symbol++) {
-        const idx = (this.numStates * symbol + state) * 3;
-        const newState = this.table[idx];
-        const newSymbol = this.table[idx + 1];
-        const action = this.table[idx + 2];
-        
-        table.push({
-          state,
-          symbol,
-          newState,
-          newSymbol,
-          action: DIRECTION_NAMES[action]
-        });
-      }
-    }
-    
-    return table;
-  }
-  
-  getActiveStates() {
-    const states = new Set();
-    for (let i = 0; i < this.heads.length; i++) {
-      states.add(this.heads[i].state);
-    }
-    return Array.from(states);
-  }
-}
 
 function renderToCanvas(canvas, prog, colors) {
   if (!canvas || !prog) return;
@@ -312,201 +90,6 @@ function renderToCanvas(canvas, prog, colors) {
   ctx.putImageData(imageData, 0, 0);
 }
 
-const ColorPalette = memo(({ palette }) => {
-  return (
-    <div className="mb-4">
-      <h2 className="text-xl font-bold mb-2">Color Palette</h2>
-      <div className="flex justify-center gap-2 flex-wrap max-w-xl">
-        {palette.map((color, i) => (
-          <div key={i} className="text-center">
-            <div
-              className="w-6 h-6 rounded-full"
-              style={{ backgroundColor: `rgb(${color[0]}, ${color[1]}, ${color[2]})` }}
-            ></div>
-            <p className="text-xs">{i}</p>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-});
-
-const TransitionTable = memo(({ transitionData }) => {
-  return (
-    <div className="mb-4 w-full max-w-4xl overflow-x-auto">
-      <h2 className="text-xl font-bold mb-2">Transition Table</h2>
-      <table className="min-w-full bg-white border border-gray-300">
-        <thead>
-          <tr className="bg-gray-200">
-            <th className="py-2 px-4 border border-gray-300">Current State</th>
-            <th className="py-2 px-4 border border-gray-300">Current Symbol</th>
-            <th className="py-2 px-4 border border-gray-300">New State</th>
-            <th className="py-2 px-4 border border-gray-300">New Symbol</th>
-            <th className="py-2 px-4 border border-gray-300">Action</th>
-          </tr>
-        </thead>
-        <tbody>
-          {transitionData.map((transition, index) => (
-            <tr key={index} className={index % 2 === 0 ? 'bg-gray-50' : 'bg-white'}>
-              <td className="py-2 px-4 border border-gray-300 text-center">{transition.state}</td>
-              <td className="py-2 px-4 border border-gray-300 text-center">{transition.symbol}</td>
-              <td className="py-2 px-4 border border-gray-300 text-center">{transition.newState}</td>
-              <td className="py-2 px-4 border border-gray-300 text-center">{transition.newSymbol}</td>
-              <td className="py-2 px-4 border border-gray-300 text-center">{transition.action}</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-});
-
-const Controls = memo(({ 
-  numStates, setNumStates,
-  numSymbols, setNumSymbols,
-  numHeads, setNumHeads,
-  headRadius, setHeadRadius,
-  speed, setSpeed,
-  seed, setSeed,
-  generateNewSeed,
-  isRunning, toggleRunning,
-  step, reset,
-  exportToJson, showTransitionTable,
-  setShowTransitionTable, fileInputRef
-}) => {
-  return (
-    <>
-      <div className="mb-4 flex gap-4 flex-wrap justify-center">
-        <div className="flex flex-col">
-          <label className="mb-1">Number of States:</label>
-          <input
-            type="number"
-            min="2"
-            max="128"
-            value={numStates}
-            onChange={(e) => setNumStates(Number(e.target.value))}
-            className="p-2 border rounded"
-          />
-        </div>
-        
-        <div className="flex flex-col">
-          <label className="mb-1">Number of Symbols:</label>
-          <input
-            type="number"
-            min="2"
-            max="128"
-            value={numSymbols}
-            onChange={(e) => setNumSymbols(Number(e.target.value))}
-            className="p-2 border rounded"
-          />
-        </div>
-        
-        <div className="flex flex-col">
-          <label className="mb-1">Number of Heads:</label>
-          <input
-            type="number"
-            min="1"
-            max="1024"
-            value={numHeads}
-            onChange={(e) => setNumHeads(Number(e.target.value))}
-            className="p-2 border rounded"
-          />
-        </div>
-        
-        <div className="flex flex-col">
-          <label className="mb-1">Head Radius:</label>
-          <input
-            type="number"
-            min="0.01"
-            max="1"
-            step="0.01"
-            value={headRadius}
-            onChange={(e) => setHeadRadius(Number(e.target.value))}
-            className="p-2 border rounded"
-          />
-        </div>
-        
-        <div className="flex flex-col">
-          <label className="mb-1">Speed (ms per step):</label>
-          <input
-            type="number"
-            min="0"
-            max="1000"
-            step="10"
-            value={speed}
-            onChange={(e) => setSpeed(Number(e.target.value))}
-            className="p-2 border rounded"
-          />
-        </div>
-        
-        <div className="flex flex-col">
-          <label className="mb-1">Seed:</label>
-          <div className="flex">
-            <input
-              type="number"
-              min="0"
-              value={seed}
-              onChange={(e) => setSeed(Number(e.target.value))}
-              className="p-2 border rounded"
-            />
-            <button
-              onClick={generateNewSeed}
-              className="ml-2 px-2 bg-gray-500 text-white rounded hover:bg-gray-600"
-              title="Generate new random seed"
-            >
-              🎲
-            </button>
-          </div>
-        </div>
-      </div>
-      
-      <div className="flex gap-4 mb-4 flex-wrap justify-center">
-        <button
-          onClick={reset}
-          className="px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600"
-        >
-          Reset
-        </button>
-        
-        <button
-          onClick={toggleRunning}
-          className={`px-4 py-2 ${isRunning ? 'bg-red-500 hover:bg-red-600' : 'bg-green-500 hover:bg-green-600'} text-white rounded`}
-        >
-          {isRunning ? 'Stop' : 'Start'}
-        </button>
-        
-        <button
-          onClick={step}
-          disabled={isRunning}
-          className={`px-4 py-2 bg-yellow-500 text-white rounded hover:bg-yellow-600 ${isRunning ? 'opacity-50 cursor-not-allowed' : ''}`}
-        >
-          Step
-        </button>
-        
-        <button
-          onClick={exportToJson}
-          className="px-4 py-2 bg-purple-500 text-white rounded hover:bg-purple-600"
-        >
-          Export JSON
-        </button>
-        
-        <button
-          onClick={() => fileInputRef.current.click()}
-          className="px-4 py-2 bg-indigo-500 text-white rounded hover:bg-indigo-600"
-        >
-          Load JSON
-        </button>
-        
-        <button
-          onClick={() => setShowTransitionTable(!showTransitionTable)}
-          className="px-4 py-2 bg-teal-500 text-white rounded hover:bg-teal-600"
-        >
-          {showTransitionTable ? 'Hide Transition Table' : 'Show Transition Table'}
-        </button>
-      </div>
-    </>
-  );
-});
 
 const TuringMachine2D = () => {
   const canvasWidth = 512;
